Check for passkeys
